Fix the warnings coming from json.hpp
Added a static cast to int in length() & size() funcs
The warning was pretty annoying.

+++ Engine/Engine/json.hpp
@@ -279,7 +279,7 @@ namespace json {
 
 		int length() const {
 			if (Type == Class::Array)
-				return Internal.List->size();
+				return static_cast<int>(Internal.List->size());
 			else
 				return -1;
 		}
@@ -292,9 +292,9 @@ namespace json {
 
 		int size() const {
 			if (Type == Class::Object)
-				return Internal.Map->size();
+				return static_cast<int>(Internal.Map->size());
 			else if (Type == Class::Array)
-				return Internal.List->size();
+				return static_cast<int>(Internal.List->size());
 			else
 				return -1;
 		}
